Introduce an Autotools-based build system
This allows installing libcangjie the way one installs most FOSS.

It passes `make distcheck`, with all important files ending up in the
tarball.

The LICENSE file was renamed COPYING so it gets included automatically
in the tarball as well.

The next step would be to build the Berkeley DB files in data/ at build
time, and removing them from Git and the tarball.