Rename the table LICENSE file
It makes no sense to call this file LICENSE, when the Public Domain is
an absence of license.

COPYING is widely used in Free Software projects, so that seems like a
good candidate for a name.

However, to avoid conflicting with the libcangjie COPYING, let's suffix
this one with .tables.