util: Comment on our utilities
If we ever grow our number of little helper functions, it
might be good idea at some point to start looking into reusing
alternatives from other libraries.

The 2 ones we have currently exist in Boost, so let's keep
track of it, in case we ever decide to switch to it in the
future.