Add a new source for the classic frequency
The current quick-classic.txt file is (most likely) under the GPL, which
might (or might not) cause us some legal issues regarding our licensing,
as we are a LGPL library.

But as it turns out, the frequency in that file had been derived from
the order of characters in the Big5 encoding.

This is a clean-room reimplementation of that frequency table, which I
hereby place in the Public Domain.

This commit also includes the script I used to build the table, which is
under the LGPLv3+, like the rest of libcangjie.