classicfreq: Be less naive about Big5 ordering
The order we were targetting is not just the order of characters in
Big5, it is slightly more complex than that.

Big5 is divided into zones. The three zones we are interested in are
called "Frequently used characters", "Graphical characters" and "Less
frequently used characters".

Inside each zone, we want to keep the order of Big5. However, zonnes
should be prioritized.

This commit now ensures that the "Frequently used characters" are always
before the other two, but doesn't prioritize "Graphical characters"
relatively to "Less frequently used characters".

This should be much closer to what the classic frequency is (i.e the
frequency on that other operating system), especially considering that
it doesn't even have all these characters we have.